instead of eagerly closing file descriptors in error conditions, behave in a unsurprising way: never close any file descriptors which have been passed in.  Provide both close_tls and close functions for terminating connections.
The function of_t now has an optional argument ?close, which by
default closes the underlying connection.

Thus, the lifecycles of file descriptors is streamlined with the
Lwt_unix interface: whenever an error occurs, the caller is
supposed to close the connection.

The only exception is accept and connect if serer_of_fd/client_of_fd
fails the TLS upgrade -- here the newly allocated file descriptor is
silently closed.  This prevents fd leakage, since the fd is not
exposed to a user.