hs_can_handle_appdata: a never-ending story?
Initially we allowed application data always after the first handshake.

Turns out, between CCS and Finished there is new crypto_context in place which
has not yet been authenticated -- bad idea to accept application data at that
point (beginning of 2015 in OCaml TLS).

The fix was to only allow application data in Established state (and block in
Tls_lwt/Tls_mirage when the user requested renegotiation) (December 2015 in
OCaml-TLS).

Renegotiation was also turned off by default when we introduced resumption (mid
October 2015): both features together (without mitigating via session hash)
allow the triple handshake.

It turns out, the server side can happily accept application data from the other
side when it just sent a HelloRequest (and waits for the ClientHello; same is
true for the client side, waiting for the ServerHello in renegotiation case
might be interleaved with application data) to let the client initiate a new
handshake.  By this commit, OCaml-TLS allows application data then.

In the end, it is a pretty academic thing anyways, since nobody uses
renegotiation with OCaml-TLS in the field.