Another attempt at the semantics of renegotiation
There is a valid use case for renegotiation, which is to require client
authentication without the client exposing their identity on the wire.

This can be either server-initiated (immediately send a HelloRequest after
completion of the initial handshake) or client-initiated (send a ClientHello
directly after the initial handshake).  While the former is more intuitive from
a system design point of view, some protocols (such as TOR) require the latter.

What is crucial is to not transmit the (plaintext) client certificate during
the initial handshake.  To get this, three optional arguments are provided by
Engine.reneg: authenticator, acceptable_cas and cert.  These modify the
configuration of the current TLS session (via Config.with_YYY).

The effectful layers (Tls_lwt and Tls_mirage) provide `reneg` with these
optional arguments as well.  Both server and client initiated renegotiation has
now (again) blocking semantics (by using `handshake_in_progress`), and `reneg`
does wait until a new handshake is performed.  Previously the server initiated
renegotiation returned immediately, which required clients to busy loop and
detect an epoch change on their own.

As it is at the moment, any APPLICATION_DATA received after HelloRequest (or
ClientHello) is sent and before Finished leads to an error.  Any
APPLICATION_DATA received before the renegotiation is initiated may be put into
`linger` (which is now by default cleared before the renegotiation is
initiated).  The reason behind this is that any data received in the old epoch
should not be confused (since it likely has other authentication credentials)
with data received in the new epoch (some applications may not care, and can
explicitly pass `~drop:false` to `reneg`).