Config/Engine/Lwt: allow a new authenticator to be passed during renegotiation
If client authentication is used with TLS, the client certificate itself is not
encrypted (neither is the server certificate).   To hide the identity of clients
(or transport secret material via a client certificate), a server-authenticated
TLS connection can be established, and immediately afterwards, a renegotiation
can be initiated.  This renegotiation is another full TLS handshake, and can
be mutually authenticated - the server can request a client certificate - and
the renegotiated handshake is fully encrypted.