Handshake_client: unbreak client renegotiation if ExtendedMasterSecret is used
In answer_server_hello_renegotiate, the session never set extended_ms to true,
which needs to be the case if both ServerHello and ClientHello contained the
extension.  The asymmetry between answer_server_hello and
answer_server_hello_renegotiate was introduced in June 2015 (0.7.0,
commit dac9e395b73321430917515d1481e15d5ec7b0a6, PR #287).

A test setup is to run
mbedtls_ssl_server2 renegotiate=1 renegotiation=1 renego_period=2 exchanges=3
and on our other side a test_client or similar where ~reneg:true is passed to
Config.client.

Instead of copy and pasting code betweent the two functions, I took the
opportunity to unify some parts into common functions (as was already partially
done with valid_cipher).