Some improvements, simplifications, and stricter types
tweaked the code a bit to use stricter types and conform more cleanly with
the RFC.

In RFC 7301, it says 'The protocol identified in the
"application_layer_protocol_negotiation" extension type in the ServerHello
SHALL be definitive for the connection, until renegotiated.'.

This means, the alpn_protocol is local to each renegotiation (instead of global
for a complete TLS session):
- move alpn_protocol from handshake_state (which is TLS-session global, such
  as protocol version that may not change during a renegotiation) to
  session_data (where all the other parameters of a negotiation are kept)

From RFC 7301 'except that the "ProtocolNameList" MUST contain exactly one
"ProtocolName".'  To ensure this invariant, Core.server_extension now contains
`` `ALPN of string `` -- and the Reader and Writer are adjusted to deal with it.
This is similar to how `` `Hostname `` works: the client_extension carries a single
string, while the server_extension must be empty (`` `Hostname `` is a constructor
of cardinality 0).  The Reader and Writer are adapted for this change.

In epoch_data, I renamed peer_alpn to alpn_protocol.  The prefix `peer_` is used
for information about the other endpoint, but alpn is (similar to ciphersuite)
negotiated once.

Handshake_client safety: if server responded with a alpn protocol, ensure it is
one we actually asked for.

Both Handshake_client and Handshake_server now include the negotiated
alpn_protocol in the session_data.

Handshake_server processes the ALPN extension in answer_client_hello_common,
which is used for every renegotiation.  Previously, the ALPN was only processed

Removed `type alpn_protocol = string`, no reason to alias the type.
in the initial handshake.  Also, `first_match` (from Utils) is reused.