tls_lwt: record io error only if state used to be active
this ensures proper Lwt_io interoperability, esp. if an Eof has been received,
which in the TLS world is replied to with a close_notify alert -- this write
can certainly fail. earlier, this resulted in an `Error state, now the `Eof
state is retained and Lwt_io can read all application data.