Improve ECC support in TLS 1.2 (RFC 8422)
The Go implementation of TLS dislikes client hello without a ECPointFormats
extension, so let's revert from #413 and re-add this extension with a single
value (uncompressed) to client hello (and server hello if client sent it, and
an ECC ciphersuite was selected).

In TLS 1.3 only settings, this change is a no-op.