Configuration: do not filter signature_algorithms based on certificate chains.
This is crucial for client authentication, especially if the server and client
certificate have different key types. Previously, if the server was using e.g.
a RSA certificate, the only signature_algorithms sent in the CertificateRequest
were RSA. Now, if the client certificate is an EC certificate, the client can't
present it, and fails.