Add tls-eio (#451)
* Clone lwt code as starting point for Eio support

Removed most of the examples for now.

* Convert Lwt test code to MDX

This MDX is based on the Lwt test_client.ml and test_server.ml files.
It runs them both together, in an Eio main loop, using Lwt_eio.

* Rename _lwt to _eio

* Move RNG initialisation out of library

Doesn't work on Eio, and also prevents users from choosing any
parameters, since future inits are ignored.

* Use Eio to create the socket

* Remove Lwt from Tls_eio

* Remove Lwt from X509_eio

* Use mock RNG for tests

Avoids depending on the (unreleased) mirage-crypto-rng-eio.

* Rename Tls_eio.Unix module

It doesn't have anything to do with Unix now.

* Code tidying

* Tidying: simplify end-of-file handling

* Simply invalid_arg reporting

* Report Eof as End_of_file, not Invalid_argument

It's (probably) not the sender's fault the connection is closed.
tls-mirage returns `Closed` in this case, while tls_lwt faised
Invalid_argument.

* Eio_main is only needed for the tests

* Temporary fix for tls-async in CI

tls-async can't be tested on OCaml 5 yet, and this prevents the CI from
testing other packages there. This annotation can be removed once a
compatible version of `core` is released.

* Use a real RNG for tests

This adds a test dependency on mirage-crypto-rng-eio.

Requested by Hannes Mehnert.

* Require dune >= 3 in all opam files

* Remove homepage from dune-project

Doesn't seem to be needed.