tls-eio: Fix end-of-file handling
I introduced this bug in f32809976cdf03c19ade77 (not in the history due
to squash merge). However, setting the state to `Eof` is not the same as
reporting end-of-file to the user, and we may still have data to report.
This can cause us to miss data at the end of the stream in some cases.