tls-eio: add fuzz tests using crowbar
The test picks two random strings, one for the client to send and one
for the server. It then runs a send and receive fiber for each end.

A dispatcher fiber then sends commands to these worker fibers. It may:

- Ask a send fiber to send some plaintext from the message.
- Ask a receive fiber to read some plaintext.
- Ask the network to transmit some number of bytes in some direction.
- Ask a send fiber to shut down the connection.
- Yield (allowing the workers to make progress).

At the end, it asks any remaining send fibers to shut down, allows all
network traffic to be transmitted, and waits for the receive fibers to
end. It then checks that the receiver got everything that was sent.

By default, this just runs a few random examples. To use afl-fuzz:

    dune runtest
    mkdir input
    echo hi > input/foo
    cp certificates/server.{key,pem} .
    afl-fuzz -m 1000 -i input -o output ./_build/default/eio/tests/fuzz.exe @@