tls-eio: allow cancelling reads
Previously, Tls would treat any exception from a read as a permanent
error, causing all future operations to fail. However, it is reasonable
to cancel a read and then try again later.

Note that the Lwt version appears to have the same problem. Possibly we
should only be catching `Eio.Io` exceptions in the first place, or maybe
none at all.

I also extended the fuzz tests to check for this case.