relax signature algorithms extension in client hello
OpenSSL, even if -tls1 or -tls1_1 are provided, adds such an extension in the
Client Hello

from RFC5246: Clients MUST NOT offer it if they are offering prior versions.
However, even if clients do offer it, the rules specified in [TLSEXT]
require servers to ignore extensions they do not understand.