tls-eio: change error when writing to a closed connection
Previously, writing to a closed connection raised `End_of_file`, but
that is misleading (it's normally used when reading from an input
stream, not when writing). This error normally occurs when the other end
shut down one side of the connection (which currently causes both
directions to be shut down).

`tls_lwt` raises `Invalid_argument` in this case, but it seems more
useful to group this with other errors where the remote end closes the
connection, so I used Eio's `Connection_reset` error here.

This was previously confusing the `copy` operation, which assumed there
was no more data to copy rather than that the write had failed, and so
didn't report an error at all in this case.

I also enabled logging in the fuzz tests by default (at warning level).
As well as ensuring that warnings are displayed, this also makes it
easy to switch to debug logging when needed.